python/requirements.txt
requests>=2.25.0
twitchAPI>=3.10.0
streamlink>=5.0.0
python-dotenv>=0.19.0
websocket-client>=1.2.1
pytest>=6.2.5
pytest-cov>=2.12.1
pillow>=9.0.0
jsonschema>=4.0.0
pytz>=2021.1
aiohttp>=3.8.0
async-timeout>=4.0.2
packaging>=21.0 